[update] : Appを修正
